+++ src/main/java/net/gunivers/gunibot/command/commands/administrator/WormHoleCommand.java
@@ -30,10 +30,10 @@ import net.gunivers.gunibot.core.lib.EmbedBuilder;
 import net.gunivers.gunibot.core.lib.EmbedBuilder.Field;
 import net.gunivers.gunibot.core.utils.BotUtils;
 import net.gunivers.gunibot.core.utils.ListUtils;
-import net.gunivers.gunibot.utils.tuple.Tuple;
-import net.gunivers.gunibot.utils.tuple.Tuple2;
-import net.gunivers.gunibot.utils.tuple.Tuple3;
-import net.gunivers.gunibot.utils.tuple.Tuple5;
+import net.gunivers.gunibot.core.utils.tuple.Tuple;
+import net.gunivers.gunibot.core.utils.tuple.Tuple2;
+import net.gunivers.gunibot.core.utils.tuple.Tuple3;
+import net.gunivers.gunibot.core.utils.tuple.Tuple5;
 import reactor.core.Disposable;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
