+++ src/main/java/net/gunivers/gunibot/core/event/CommandIssuedListener.java
@@ -16,8 +16,8 @@ import net.gunivers.gunibot.Main;
 import net.gunivers.gunibot.command.permissions.Permission;
 import net.gunivers.gunibot.core.command.Command;
 import net.gunivers.gunibot.core.datas.DataGuild;
-import net.gunivers.gunibot.utils.tuple.Tuple;
-import net.gunivers.gunibot.utils.tuple.Tuple2;
+import net.gunivers.gunibot.core.utils.tuple.Tuple;
+import net.gunivers.gunibot.core.utils.tuple.Tuple2;
 
 public class CommandIssuedListener extends Events<MessageCreateEvent>
 {
