+++ src/main/java/net/gunivers/gunibot/core/utils/tuple/Tuple1.java
@@ -1,4 +1,4 @@
-package net.gunivers.gunibot.utils.tuple;
+package net.gunivers.gunibot.core.utils.tuple;
 
 import java.io.Serializable;
